+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 - Add `vault lookup-token` (#77)
 - Remove `client.get_all()` (after deprecation period)
+- Add `vault ln` and `--follow` features in client and cli (#65)
 
 0.6.0 (2019-15-06)
 ------------------